[FIX] mail: no seen indicator in channels of type 'channel'
Before this commit, the seen indicator were displayed in the
channels of type 'channel'. This was not intentional, and
actually caused a UI issue where the seen indicator was
visible on all messages before the last message relevant
to display the seen indicator.

Step to reproduce:

- As Mitchell Admin, post 2 messages in `#general`
- As Marc Demo, post a single message in `#general`
=> Mitchell Admin sees the seen indicator on both of his messages,
instead of 1 (feature enabled) or 0 (feature disabled).

The 2 indicators happen because the channels is considered as not
elligible for seen indicator feature. Because of this, field
`Thread.lastSelfMessageSeenByEveryone` is unset, which also sets
`Message.isMessagePreviousToLastSelfMessageSeenByEveryone` to `false`
and thus makes all fields elligible to display the seen indicator.

This commit fixes the issue by showing the seen indicator only when
the thread of channel supports seen indicator, that is chat channels.

=====

This enterprise commit adapts whatsapp code to properly flag whatsapp
channels as supporting the seen feature. Tests were passing due to
bug on `'channel'` channels.

closes odoo/enterprise#62750

Related: odoo/odoo#165883
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>