[IMP] documents_account: Add selection of journal in Document Action 'import bank statement'
Backport of: https://github.com/odoo/enterprise/pull/46482/commits/d2a95d7ffa3ef2976172ed17eaf93056ea0f8c6c

For now we didn't chose the bank journal because we assume that only
coda file where push like this, and in the coda file we can find in
which journal to send it. But Coda is limited to BE and lot of user
import CSV file for bank.

Add the selection of the journal (not mandatory) for Document action,
and inject CSV in the journal selected. If no journal selected inject
the CSV in the first one of the list. If it's a coda file, keep as now
and don't take care of the journal selected.

task-3932055

closes odoo/enterprise#65041

X-original-commit: 58288405690ef5f9b93bedb350fe6a353b49543a
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Muhamed Abdalla (muab) <muab@odoo.com>