[FIX] documents: add parent workspace name in sub workspace name
Steps to reproduce:
-------------------
- Install `Documents` module
- Go to Documents -> Configuration -> Workspaces
- Create a workspace with a sub workspace
- Go to list view

Issue:
------
The display name of the sub workspace is not correct;
It should also add the parent workspace name in the display name.

Cause:
------
The following commit added a second `_compute_display_name` method
that override the flow that add the parent workspace to the sub
workspace name.
https://github.com/odoo/enterprise/commit/ca7efce8f7bdb497f39a8438b5c266a7861d9a92

Solution:
---------
Merge the 2 `_compute_display_name`.

opw-3835887

closes odoo/enterprise#63082

X-original-commit: a74e9f3c9348c633ab8628cd15ccca1defb2280a
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>