[FIX] appointment_hr: update query count after res.users change
This commits updates three query counter checks adding three extra queries in order to
adapt them to the addition of 1. a new sub query in 'calendar.event' which refers to
the 'res.users.settings' table, 2. the fetching of the Default User Template in the
'res.users' model and 3. the search of the current user's 'res_users_settings_id'
configuration.

related-to: odoo/odoo#162467
task-3850837

closes odoo/enterprise#62767

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>