[REV] account_accountant: revert deferred date in pdf
In this commit https://github.com/odoo/enterprise/commit/84e0edcfc88c3361e1e042e80c82f6c4b950c738,
we added the deferred_start_date and end_date in the pdf. This change is applied
for everyone and it shouldn't be.

A next task will be created to assess this problem but in the meantime we will
revert the commit

no task-id

closes odoo/enterprise#62775

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>