[FIX] l10n_in_reports_*: rectify calculation for GSTR-1 Table-14(a)
Before this commit:
Table 14(a) in the GSTR-1 report was calculating incorrect
entries.

After this commit:
Table 14(a) in the GSTR-1 now calculates correct entries.

Technical Reason:
Previously, the calculations were based on incorrect tax_tag_ids and subformula.
Now, the calculations are made with the correct tax_tag_ids and subformula, and
cess has also been added.

task-3904018

closes odoo/enterprise#63227

X-original-commit: 166fb40407db4f272b7d16709b06455e2b5fc8a1
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Harsh Trivedi (hatr) <hatr@odoo.com>