[FIX] l10n_in_reports_gstr: Fix GSTR1 JSON Error by Removing unsupported supeco
While filing GSTR-1, we encountered the following error:

```
[0]JSON validation failed due to - #: extraneous key [supeco] is not permitted
```

In this commit, we have resolved the error mentioned above. For various reasons,
the government API has not yet implemented GSTR-1 tables 14 and 15 in production
though they are available in the sandbox environment. As a solution, it is best
not to include supeco in the JSON until the government API releases supeco into
production.

closes odoo/enterprise#62835

X-original-commit: 8dcd70277e7aec82392897430513e070b7da0c08
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Harsh Modi (hamo) <hamo@odoo.com>