[FIX] mrp_workorder: use correct type for context
**Current behavior:**
Creating a manufacturing order for a product which is tracked
via lots can cause a trace back when trying to register the lot
number in the shop floor as part of the registering production
flow.

**Expected behavior:**
No trace back.

**Steps to reproduce:**
1. Create a product which is tracked with lots and a component

2. Create an MO to manufacture the lot-tracked product and
     confirm it

3. Open the shop floor, click on the MO line for the product to
     add the lot number and check it off -> trace back

**Cause of the issue:**
The context prop is a 'stringified' object instead of the
owl-enforced 'object' type.

`MrpRegisterProductionDialog` extends
`MrpQualityCheckConfirmationDialog`, where we can see context is
not stringified.

**Fix:**
Don't stringify the context upon creation of the confirmation
dialog object.

opw-3876368

closes odoo/enterprise#63474

X-original-commit: eee7b6a3965fa154933604c0833ff195c3114588
Signed-off-by: Steve Van Essche <svs@odoo.com>