[IMP] documents{_spreadsheet}: improve burger menu
Improve the burger menu ui:
- reducing the labels size.
- regrouping all the actions related to a document addition
(upload, request, link, spreadsheet) under a "New" button dropdown.
- Get the "share" button out of the dropdown, styled secondary.
- Add the possibility to create a workspace from the "New" dropdown.
Only available for the users with the group "group_documents_manager".
The new workspace will have the current folder as parent. If the current
folder is "All" or if there's no current folder it'll be placed to the
default location (under the "Internals" folder).

Technical:
- As the "groups" attribute doesn't work in the ControlPanel xml,
retrieving the user groups from the js.
- Removing the test named "when no sharable workspace is present, check
the visibility of dropdown button inside 'All' workspace" as the
dropdown button is now renamed "New" and is always enabled.
- As the _reloadSearchModel method was present at 3 different places
(DocumentsSearchModel, DocumentsSearchPanel and DocumentsInspector),
removing the redundancy by only declaring it in the DocumentsSearchModel.

Task-3889854

closes odoo/enterprise#61653

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>