[IMP] planning: add tests for the planning's front end
Currently there are no tests for the planning's front end.
This means that changes in the code could break the front
end's functionality without anyone noticing.

Unfortunately, with the way that the front end is set-up
there is no formal way to test the front-end like the
back-end.

This commit adds three tours which should make sure that the
main functionalities of the front end are tested for future
potential changes.

task-3800770

closes odoo/enterprise#63095

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>