[IMP] l10n_mx_edi: allow linking CFDI invoice to existing vendor bill
Previously, you could only link a vendor bill with a CFDI invoice file
by creating the bill directly from it. In a lot of cases, Odoo is unable
to match bills created from the CFDI file with the corresponding
purchase order for various reasons (eg: different line items, different
prices, etc), and it is currently impossible to link them by hand.

This commit makes it so that when you attach a CFDI invoice to a vendor
bill, if that bill does not already have an associated CFDI invoice, it
becomes associated with that CFDI invoice, and the fiscal folio is
automatically filled in from the document. If there are lines on the
vendor bill, lines from the CFDI invoice are ignored, if there are none
they are filled in as normal.

task-3731034

closes odoo/enterprise#64794

X-original-commit: b31d648dd89a3589acee24a652a7aca59efef017
Related: odoo/odoo#169631
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>