[FIX] l10n_es_reports: improve "mod347 report" search domain perf
The model 347 report is adding a partner exclusion domain in the search
call.
It is done by firstly retrieving those partner IDs to then add that list
of IDs into a `not in` domain leaf.
The issue is that the list of IDs can be gigantic, leading to
performance issues.

Before commit [1], it was a blocking issue as if the list was too long
(100,000 IDs raise the issue), it would simply make the code crash when
loading the report.
Commit [1] then simply bypassed the issue by passing the domain as
`forced_domain` so it would not go through the `literal_eval()` call
(which is the one crashing when the list is too long).

But even after that fix, we still provide the list of IDs into the
domain and so ultimately in the SQL request, which can still be an issue
if there are too many.

This commit improves that by passing the SQL subquery (the one fetching
the partners to exclude) directly into the domain, so the SQL query
won't contain any list of IDs but just a sub-select.
It does have better performance within PostgreSQL for big lists.

Note that the `SQL()` query is surrounded by parentheses on purpose, as
it doesn't work otherwise. This might be improved at the ORM level in
the future with [2].

[1]: https://github.com/odoo/enterprise/commit/490bd2c0e067d3105e30c8b916d5873908a0805f
[2]: https://github.com/odoo/odoo/pull/163560

task-3895983

closes odoo/enterprise#66855

X-original-commit: 1179ea6e5e55e5ec6e4756773e7478900c260003
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Romain Derie (rde) <rde@odoo.com>