[FIX] helpdesk: create ticket with email cc field
Steps to reproduce:
- Install `website_helpdesk`
- Edit the help form to add a field of type 'Email CC'
- Fill the form and for the 'Email CC' put a random mail

Issues:
The ticket is created but doesn't have any subscribers.

The reason is that since the supplied mail in `mail_cc`
is not related to any internal user test we will hit
the continue in the `create` method.

https://github.com/odoo/enterprise/blob/ee506c2ab3e7b39831aed0af2142f8f7e4ba02ba/helpdesk/models/helpdesk_ticket.py#L508-L509

Because of this continue we will skip the code in charge of subscribing
the partner to the ticket.

https://github.com/odoo/enterprise/blob/ee506c2ab3e7b39831aed0af2142f8f7e4ba02ba/helpdesk/models/helpdesk_ticket.py#L512-L513

opw-3922047

closes odoo/enterprise#63525

X-original-commit: 95961d2d6020fed29ac6eceffac7e24f90685a2c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mattis Megevand (mmeg) <mmeg@odoo.com>