[FIX] l10n_nl_reports_sbr_icp: Add missing invisible fields in the view, required by computed field
The compute function _compute_sending_conditions needs date_from and date_to fields as its dependencies to compute whether or not to show the "Send" button on the ICP report SBR wizard. Without them, the field does not get properly recomptued, and the button isn't shown, unless the "Is Test" checkbox gets ticked by the user.

closes odoo/enterprise#62924

X-original-commit: 913e4fe4a22eaadee26a37f184625de68acbd1dd
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>