[FIX] mrp_maintenance: incorrect datetime parsing with trailing milliseconds
Before this commit:
The schedule date is converted into a string in SQL and then parsed in Python
using `Datetime.to_string`. However, it does not support parsing datetime with
trailing milliseconds, such as '2024-05-22 10:09:45.5'.

After this commit:
Do not convert the schedule date to a string but put it in a PostgreSQL Array
instead. Psycopg2 will convert it into a datetime object automatically,
similar to the ORM's behaviour.

opw-3932891

closes odoo/enterprise#67469

closes odoo/enterprise#67593

X-original-commit: d96549f56fc2093e573fb8918dddc1f3d89c3c04
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Walravens Mathieu (wama) <wama@odoo.com>