[ADD] web_studio: implement a class that diffs xml trees with keys
Previous state of the art
-------------------------
The report editor saved its changes by bundling all the new arch of a view into
one big inherited view that overriden every inherited view before it.
Editing client views (not qweb) was a bit different. The `ir.ui.view:normalize` method tried
to compute a minimal inherited view into a structured xml arch.

The `normalize` method was not applicable to the use case of the report editor as it was
tailored for views (references to form, notebook, fields etc....).
Moreover, it was not very maintainable, lengthy, complicated and at the end of the day rather weak.
Namely, since the diff was basically textual, determining if a node is the same as another one
is very tricky and very error prone.

Rationale of this commit
------------------------
In an effort to normalize and synthesize the diff made by a user in a report via the report editor
this commit proposes a new tool to build an arch that contains all those changes.
It ambitions to create a tool that can handle diffing both the qweb views and the client views.
It also ambitions to make the code stronger, more understandable and more maintainable.

Some diffing tools already exist in the eco-system, for example [xmldiff](https://pypi.org/project/xmldiff/) is promising but its API is unstable by the author's standard. Generally all these tools seem to lack the maturity to be included as a Odoo dependency.
Moreover, the Odoo xpath DSL is not complete enough to support a simple translation from a diff to an xpath arch.
In addition, the use case we want to handle (generate an xpath of what happened to a tree) is simple enough and doesn't require
the diff tool to be able to handle arbitrary trees.

Implementation details
----------------------
The diffing tool (KeyedXmlDiffer) relies mandatorily on the nodes to have a unique id in both trees. A node in the
"new" tree having the same id of a node in the "old" tree are considered the same, and will be diffed as such.
Hence, the recommended use of this tool is to:
- Have a tree that will be edited
- Assign unique id to each Element
- Have the user make some changes in this tree, but keep a copy of it somewhere.
- Now, we have two trees: the original and a copy on which some changes occured
- Compare them with KeyedXmlDiffer:diff_xpath. This method returns an arch containing
xpaths, that when executed by the `apply_inheritance_spec` function on the old tree
should output the new tree.

It supports:
- moving, removing nodes
- altering the order of a node's children
- altering texts
- modifying attributes of a node

It doesn't support:
- changing the tag name of a node
- anything else than etree.Element, in particular, comments and their tail will be ignored

task-3790609

Part-of: odoo/enterprise#62955
Related: odoo/odoo#166300
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>