[FIX] l10n_account_customer_statements: make customer statements from studio app not crash
When accessing customer statement from contact page in studio an error is raised.

Steps to reproduce:
- Install Studio, Contact apps and the l10n_account_customer_statements module.
- Open studio and go to Reports
- When clicking on the Customer Statements report an error is raised

Cause:
Studio uses the partner res.partner(0,) to display the report in edit mode, but the method _get_report_values is always searching for the partner data to fill the report. The error is raised because res.partner(0,) does not reference any partner.

Solution:
Modify _get_report_values to search for partner data only if the partner id is different from 0.

opw-3920298

closes odoo/enterprise#63134

X-original-commit: f25a738e17a54e7b3a012655a0f73083bfa68786
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>