[IMP] social_crm: handle image and fix links
Purpose
=======
Display the post/comment image(s) in the related generated lead description.
Also fixing a missing spacing around the social media link in the lead description.

Specifications
==============
When generating a lead from a post/comment containing an image,
only the post/comment message content is displayed in the lead description.
Adding the display of the post/comment image(s) if there is.

Task-3834228

closes odoo/enterprise#59622

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>