[FIX] account_asset: decrease in future has weird amounts.
To reproduce:
- Create an asset linear yearly.
- Reevaluate it next month, while decreasing the residual amount.
=> The amounts don't make sense, the decrease move has been deleted.
   Idem for the move before the reeval.

The issue is that we delete all the moves that are draft.
But if a move (like the decrease move) is created after the current day,
it is in auto-post, and so, in draft.

To correct that, we prevent their deletion and compute the right residual amount, with the draft moves remaining.

Another issue was found, related to already depreciated assets.
Create an asset with 10000 as original value, 4000 as Import amount
and 6000 as salvage value.
Confirm it (no moves should be created)
Sell or dispose it.

=> It creates a move before disposal while there shouldn't be any

The issue was that reevaluation didn't work with already depreciated assets and with no moves before the reevaluation.
Issue corrected with giving the right residual amount to compute the move before the pause

opw-3945341
task-3444716

closes odoo/enterprise#65192

X-original-commit: bed42d03fa86e98cb4421c027cf3d20123ab5c94
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>