[IMP] approvals: Dashboard revamp
This commit revamps the dashboard of the approvals
module. The dashboard give now the possibility to
archive/unarchive categories in addition to the
possibility to configure them. The menu permitting
to create a new approval request has been moved to
as the main button of the dashboard.
All these changes are made to make the dashboard
more user-friendly and to make it more consistent
with the rest of the module.

task-3634015

closes odoo/enterprise#62971

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>