[IMP] web_studio: use KeyedXmlDiffer for report_editor
This commit makes use of the newly introduced KeyedXmlDiffer to save the changes made by a user in a given report's view
as an inherited view which *has to come last* in the inheritance hierarchy.

The main goal is to ease the upgrade process by producing one single arch, easily readable and
on which xpaths are minimalist and somewhat self-explanatory.
With this commit, the developper or partner with knowledge of both QWeb views and Odoo's xpath DSL should be able
to quickly mentally represent the changes that were made, adapt them or reproduce them on a migrating database.

This commit is limited in its reach as it would not apply for diffing two trees we know nothing about.
This is because in Odoo there is no reliable way to identify a node. In edition mode, most attributes could change
and in the upgrade process, everything may be expected to change.

task-3790609

closes odoo/enterprise#62955

Related: odoo/odoo#166300
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>