[FIX] knowledge: toolbar in mobile overlaps article
This commit fixes an issue with the editor toolbar in mobile mode.
The toolbar overlaps some elements inside a knowledge article, e.g. separators
and embedded views. This is because in mobile the toolbar has a z-index auto
which means that it has the same z-index as the editor which can be lower than
some elements with an index set.
Before it wasn't an issue as the toolbar wasn't sticky so it disappeared when
you scrolled on mobile.
But the `position: sticky;` forces us to take the z-index into account.

To fix this, the css rule is removed so that the toolbar uses its original index
which places it back on top of the other blocks.

task-3939911

closes odoo/enterprise#63553

X-original-commit: 6c6b6dc9690f79d09e1528c402f79fc27f8cc358
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>