[FIX] voip: prevent crash when null sender track
__Current behavior before commit:__
When there is an incoming call, a VoIP provider may give a session
description containing an audio sender and a video sender even if the
video track is empty.

In such case, `track` may be `null`. Therefore the page will crash when
we try to set `track.enabled`.

__Description of the fix:__
Check if each track is not `null` before setting its `enabled` property.

opw-3833811

closes odoo/enterprise#62980

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>