[FIX] room: remove background not found errors in qunit tests
Purpose:
--------
The frontend room view uses the "background-image" property to add a
background image if one has been set on the room.
This causes several 404 errors in qunit tests because the test room does
not exist, resulting in a "room not found" exception when the browser
tries to get the background image.

This commit patches the view in the tests to remove the url part of the
background-image property, so that the background image is not fetched
anymore.

Task-3943099

closes odoo/enterprise#64055

X-original-commit: 223aa8078b54c50f017016f6e39598377af784ab
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Adrien Schoffeniels (adsc) <adsc@odoo.com>