[FIX] stock_barcode: changeDestinationLocation
Before this commit, the code in `BarcodePickingModel`
`_processLocationDestination` can call multiple times
`changeDestinationLocation` for the same line, which is not wanted.

It was because the `selectedLine` can be two times in the `lines` array
because it's the line we take by default and it is usually in
`previousScannedLines` too and we take all lines from this array when the
scan destination config is on "optional".

X-original-commit: ff1f318c439b5bfcf7751c6c48b3a0e5f3d4fcc5
Part-of: odoo/enterprise#63172