[FIX] stock_barcode: changeDestinationLocation
Before this commit, the code in `BarcodePickingModel`
`_processLocationDestination` can call multiple times
`changeDestinationLocation` for the same line, which is not wanted.

It was because the `selectedLine` can be two times in the `lines` array
because it's the line we take by default and it is usually in
`previousScannedLines` too and we take all lines from this array when the
scan destination config is on "optional".

Part-of: odoo/enterprise#62733