[FIX] website_generator: simplify website name generation
Before, we had a naive approach to getting and setting the domain
of the website as the website name. This caused bugs on websites
such as localhost for example. Now the process is simplified and we
take the whole netloc of the URL, without the 'www.' infront.
E.g. www.odoo.co.uk/page -> odoo.co.uk or www.odoo.com -> odoo.com

closes odoo/enterprise#63226

X-original-commit: bbffbb6f33fece2d24080299088493f8027542fd
Signed-off-by: Ivan Klapka (ivkl) <ivkl@odoo.com>