[FIX] account_bank_statement_import_camt: use `<AddtlNtryInf>` as fallback for `payment_ref`
Currently, when importing a CAMT file, the system fails to use
`<AddtlNtryInf>` as a fallback for the payment reference.

Steps to Reproduce
------------------
1. Install `account_bank_statement_import_camt`.
2. Import a CAMT file with a statement that includes `<TxDtls>` but no
   `<RmtInf>`, and has `<AddtlNtryInf>` defined.

Expected Result: The imported statement's payment reference should
contain the value of `<AddtlNtryInf>`.

Actual Result: The imported statement's payment reference contains
`/` instead.

Cause
-----
The system attempts to use `<AddtlNtryInf>` as a fallback for
`payment_ref`. However, it looks in the wrong place. It searches inside
`<TxDtls>`, whereas `<AddtlNtryInf>` is actually a child of `<Ntry>`.

opw-3878785

closes odoo/enterprise#63112

X-original-commit: b23454989373e4b5981be2f28cf9c9b03111ac26
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>