[FIX] stock_barcode: split not picked line
Issue introduced by [1]

How to reproduce:
- Create a new product, set a barcode and set some qty in stock (e.g.: 2
  qty);
- Create and confirm a delivery for this product but with more qty than
  in stock (e.g.: 4);
- Open the delivery in the Barcode app, then click on the back button,
  return in the Inventory app and open the delivery in the back end
=> There is now 2 moves for the product : one with the reserved qty, one
   with the non-reserved qty.

The issue was: we split incomplete moves when leaving an operation from
the Barcode app to not lose the reserved qty (check [1] for more
information) but we do that also for not picked move which is not
needed.
This fix avoids not picked moves for the split.

[1] odoo/enterprise#60452

task-3869731

closes odoo/enterprise#67142

X-original-commit: 794f029222f4275c11f95e962cc5a59210f1df2f
Signed-off-by: Steve Van Essche <svs@odoo.com>