[IMP] documents: refactor the sharing flow
Purpose
=======
Avoid creating share records prematurely.

Share popup
-----------
We always create a `document.share`, and if we don't use it, we
manually make an unlink RPC call to remove the temporary record...

This is very inefficient because we make an RPC call to create the record,
one to get the action, one to remove the record if we cancel the modal, etc.

Instead, we can just not create the record at all, and refactor the
code in that direction (so the share URL needs to be just based on
the token, not on the ID, since we need to copy the URL before the
record is created).

Spreadsheet
-----------
When we share a spreadsheet, we make a snapshot of it. This snapshot
is saved in the model `documents.shared.spreadsheet`, and it can not
be done in python. Before it used the fact that the share was created
to not make an RPC call, but it can not be used anymore.

We can still avoid making an RPC call by creating a new non-stored field,
with an inverse, that way the performance remain the same.

So, we have a lot of code with RPC call to create those snapshot records,

Task-3924439

closes odoo/enterprise#63153

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>