[IMP] approvals: add automated actions menu item and add predefined triggers
In the scope of approvals, Automations are something
that should be really put in front cause the power
of the app is to adapt to each flow that's not
already defined in Odoo. So, the menu to add a new
one should be more visible.

Thus, this commit adds a menuitem in the approvals
app `Approvals app->Configuration->Automated Actions`
to ease the access to automation rules.
Also, this commit adds predefined triggers such as
`on archived` and `on unarchived` to ease their
access and usage.

task-3872042

closes odoo/enterprise#63274

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>