[FIX] stock_barcode{,_picking_batch}: sync dest id
**Current behavior:**
If we have a picking batch with some pickings all for the same
product, while enforcing mandatory scans on source location,
products, and destination location, there is some unexpected
behavior if a different destination barcode is scanned after the
similar-product pickings have all been scanned.

**Expected behavior:**
With this specific configuration, pickings with the same product
should all be changed when a different destination is scanned.

**Steps to reproduce:**
1. In the Internal Transfers picking type, configure the barcode
     settings like so:
     `Source Location: Mandatory Scan`
     `Product: True`
     `Destination Location: After each product`

2. Create some new internal location (e.g., .../Stock/shelf)

3. Create a new batch picking with 2 internal transfer pickings
     for the same product

4. Go to the batch operation barcode menu, select the created
     batch picking

5. Scan the source location, scan the first product code and
     then manually add the second picking quantity with the add
     quantity button, scan the newly created internal location

6. Observe that the last picking destination location has
     changed and the first has not- furthermore trying to change
     the second's destination location via scan does not work
     (manual changes via the edit button/form still possible)

**Cause of the issue:**
The method of changing destination does not account for this
setup, so there is no specific handling for it. Only the
selected line will be modified by a new destination scan.

**Fix:**
Extract the line collection function and override it in the
batch barcode module to handle this particular case.

opw-3733870

closes odoo/enterprise#62201

Signed-off-by: Steve Van Essche <svs@odoo.com>