[FIX] helpdesk_timesheet: don't force UTC time on timesheets
Versions
--------
- 15.0e+

Steps
-----
1. If current UTC time is before 11, set user tz to Pacific/Niue;
2. if current UTC time is after 11, set user tz to Pacific/Kiritimati;
3. go to Helpdesk;
4. enable timesheets for a helpdesk team;
5. open a ticket of that helpdesk team;
6. start timer;
7. stop timer;
8. create timesheet.

Issue
-----
The date of the timesheet is set to the UTC date instead of the user
date, which is one day off.

Cause
-----
The wizard creating timesheets adds a `Datetime.now()` value to `date`,
which defaults to UTC time.

Solution
--------
Remove this line, allowing timesheets to be created with the default
value, which does get adjusted for user timezone.

opw-3834908

closes odoo/enterprise#63774

X-original-commit: 7640860f415028d2d516a2df5e4bc26d9d6dd3b4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>