[IMP] l10n_ma_reports: Add ice number to invoicing
Get rid of the l10n_ma_ice and depend on company_registry as both should be the same
And Add the ICE number to the invoice report. it's now part of the company_details of the invoice report if and only if
it's a moroccan company

Adding migration script to replace 'l10n_ma_ice' in 'res.partner' with 'company_registry'

Reason: The ICE (Identifiant Commun de l'Entreprise) is an identification number assigned to businesses and legal entities
for various administrative and legal purposes in Morocco. If the partner has one, it must be indicated on the invoice.

--> Important Note: Fix a small bug in the 'l10n_ma_customs_vat' invisible condition of l10n_ma_reports partner view

Task-3877546

closes odoo/enterprise#63083

Related: odoo/odoo#166531
Related: odoo/upgrade#6069
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>