[IMP] knowledge: enable no-content helper on embedded views
This commit will enable the no-content helper of the embedded views of
Knowledge. When the embedded views are empty, they will now show a
no-content helper indicating to the user that there is no record to
display. The no-content helper will provide guidance to the users and
help them creating their first record.

task-3381729

Part-of: odoo/enterprise#51284