[IMP] {website_}appointment: improve confirmation page
Purpose
=======
Improve the appointment type front-end confirmation page.

Specification
=============
- Removing the "Schedule another" button and set a "<- All Appointment"
link in the top left.
- Changing the "Cancel" button by a sentence at the end of the page.
The organizer extra confirmation message displayed underneath.
- Moving the calendar buttons (outlook and google) under the date info.
- Removing the green banner confirming the appointment. Instead of that,
setting "Appointment Scheduled" as title with a green check icon.
Propagating the registration page "Back to edit mode" info banner on the
confirmation page to keep the link to the back-end.
- Replacing the long videocall link by a short sentence (i.e. Join with Odoo Discuss)
and copy the link when clicking on the provider's name.
- When the appointment type has the "Manual Confirmation" enabled. The title should
be "Appointment Reserved" with a thumbs up icon colored as "info".
- Displaying a danger alert if the user tries to cancel the event when the
pre-booking time is passed.
- Rewording the "Table Booking" appointment type by "Table" to prevent duplicating
the "Booking" word on the confirmation page.

Technical:
Removing the click on the "Back to edit mode" link in the website_appointment tests
as the alert link has been removed from the website confirmation page.

Task-3877550

Part-of: odoo/enterprise#61350