[IMP] l10n_au_hr_payroll*: Adds a payment register method to Payslip batch
This commit allows the payment for a payslip batch to be registered
directly from a batch. And it generates a account.payment.batch record
which can be used to create an ABA file directly from the batch payment.

It also removes the redundant code and uses l10n_au_aba to generate
Aba file.

task #3855817

closes odoo/enterprise#62162

Related: odoo/upgrade#5990
Related: odoo/odoo#164870
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>