[FIX] pos_preparation_display: fix display category computiation
Currently, when a user has 2 pos preparation displays, one set on some
categories the other on all, if you add on the order a product that is not
in the categories of the first display, the order will show no changes.

Steps to reproduce:
-------------------
* Go to the **Kitchen display** App
* Have two displays created
  * Display 1: `Product Categories`: Food
  * Display 2: `Product Categories`: All (Leave blank)
* Go to the **Point of Sale** App
* Open restaurant session
* Select a table and add a drink to the order
> Observations: The order button is not highlighted.

Why the fix:
------------
Before the refactoring, the categories loaded for each display would use
the function `_get_pos_category_ids()`. This function would return a list
of all possible categories (ex: `[1, 2, 3, 4, 5]`) when the setting was set
on all categories.

After the refactoring, the categories loaded correspond to the setting of
the display. Thus, in the case of a display on all categories it would return
`[]`.

At the end, we would end up concatenating stuff like `([5], [])` instead of
`([5], [1, 2, 3, 4, 5])`. Only displays with categories set would be taken
into account.

opw-3923383

closes odoo/enterprise#64302

X-original-commit: 71d861df0acaa1190cab50d612ec6d738503b45a
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>