[IMP] l10n_br_edi: mark Odoo invoices as non-legal ones with watermark
The official invoice will be retrieved via Avatax. Adding the official
PDF in the portal is a bit complicated, so we instead we add a
watermark to the standard Odoo one. We also add a message in the
portal (/my/invoices/...) to inform the user that the official
documents can be downloaded by clicking the "Download" button on the
left.

task-3719527

closes odoo/enterprise#63946

Signed-off-by: Josse Colpaert <jco@odoo.com>