[FIX] planning: tour planning_split_shift_week
Before this commit, the tour planning_split_shift_week would crash when
the current day is the beginning of the current week. This happened
because when today is the first day of the week, the selector used in
the step "Select first day of the current week" would be the start of
the previous week. Thus the final date range produced, would contain
two weeks instead of one, making the tour timeout (waitUntil condition
not met).

closes odoo/enterprise#63263

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>