[FIX] spreadsheet_edition: add missing transformation
Before this commit, deleting a list and updating its domain concurrently
would lead to a crash. This was due to the fact that the transformation
of `UPDATE_ODOO_LIST_DOMAIN` was missing.

This commit adds the missing transformation.
This commit also fix a test that checked the number of pivots instead
of the number of lists.

closes odoo/enterprise#63479

Task: 3908657
X-original-commit: 33e65451a5a259346728ffc3af0abd68b3ce3ef0
Related: odoo/odoo#167290
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>