[FIX] spreadsheet_edition: add missing transformation
Before this commit, deleting a list and updating its domain concurrently
would lead to a crash. This was due to the fact that the transformation
of `UPDATE_ODOO_LIST_DOMAIN` was missing.

This commit adds the missing transformation.
This commit also fix a test that checked the number of pivots instead
of the number of lists.

closes odoo/enterprise#63427

Task: 3908657
X-original-commit: 61f84963b43a9e57c8c0fbb0b16ae9ffd415081f
Related: odoo/odoo#167177
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>